[Format] Introduce shamformat sublib
Why you may ask ? Well it was burried in shambase & is independant of
everything else so make sense to split it right ?
