+++ src/CMakeLists.txt
@@ -6,6 +6,7 @@
 # ~~~
 
 set(as_subproject
+    shamformat
     shamunits
     shambase
     shamcomm
@@ -35,6 +36,7 @@ macro(find_package)
     endif()
 endmacro()
 
+add_subdirectory(shamformat)
 add_subdirectory(shamunits)
 add_subdirectory(shambase)
 add_subdirectory(shambindings)
