+++ src/shambase/CMakeLists.txt
@@ -16,7 +16,6 @@ set(Sources
     src/print.cpp
     src/popen.cpp
     src/term_colors.cpp
-    src/format.cpp
     src/profiling/chrome.cpp
     src/profiling/profiling.cpp
     src/logs/reformat_message.cpp
@@ -71,3 +70,4 @@ if(SHAMROCK_USE_NVTX)
 endif()
 
 target_link_libraries(shambase PUBLIC fmt::fmt-header-only)
+target_link_libraries(shambase PUBLIC shamformat)
